Main Project/Backend/requirements.txt
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
decorator==5.2.1
Flask==3.1.3
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
MarkupSafe==3.0.3
msgpack==1.1.2
numba==0.65.0
numpy==2.4.4
oauth2client==4.1.3
oauthlib==3.3.1
packaging==26.1
platformdirs==4.9.6
pooch==1.9.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
